Add Protagonist!
Also make a few other small changes.
  * Bytes are now floats in the Vertex constructor.
  * Recolored some stuff for clarity.
  * Removed trailing whitespace.

+++ src/main/scala/Vertex.scala
@@ -5,14 +5,14 @@ import java.nio.ByteBuffer
 case class Vertex(
   x: Float, y: Float, z: Float,
   nx: Float, ny: Float, nz: Float,
-  r: Byte, g: Byte, b: Byte,
-  s: Float, t: Float) 
+  r: Float, g: Float, b: Float,
+  s: Float, t: Float)
 {
   def insertIntoBuf(vBuf: ByteBuffer) = {
     vBuf
       .putFloat(x).putFloat(y).putFloat(z)
       .putFloat(nx).putFloat(ny).putFloat(nz)
-      .put(r).put(g).put(b)
+      .put(r.toByte).put(g.toByte).put(b.toByte)
       .putFloat(s).putFloat(t)
   }
 }
@@ -22,11 +22,11 @@ object Vertex {
   val norSize = java.lang.Float.SIZE/8 * 3
   val colSize = java.lang.Byte.SIZE/8 * 3
   val texSize = java.lang.Float.SIZE/8 * 2
-  
+
   val posOffset = 0
   val norOffset = posSize
   val colOffset = norOffset + norSize
   val texOffset = colOffset + colSize
-  
+
   val strideSize = texOffset + texSize
 }
